requirements-tpu.txt
--find-links https://storage.googleapis.com/jax-releases/libtpu_releases.html
jax[tpu]
flax
optax
clu
pyyaml
datasets
Pillow
grain
git+https://github.com/deepmind/jmp
